fix Errorhandler and fix renameing file and folder

+++ Views/StartView.xaml.cs
@@ -5,6 +5,7 @@ using RetroGameHandler.Helpers;
 using RetroGameHandler.models;
 using RetroGameHandler.ViewModels;
 using System;
+using System.Text.RegularExpressions;
 using System.Threading;
 using System.Windows;
 using System.Windows.Controls;
@@ -69,7 +70,7 @@ namespace RetroGameHandler.Views
             {
                 var textBlock = (TextBlock)sender;
                 textBlock.Cursor = Cursors.Wait;
-                 var ftpItmListM = (FtpListItemModel)textBlock.DataContext;
+                var ftpItmListM = (FtpListItemModel)textBlock.DataContext;
                 if (ftpItmListM.IsImage)
                 {
                     ftpItmListM.GetImage();
@@ -87,10 +88,13 @@ namespace RetroGameHandler.Views
                 }
                 DirFileInfo.DataContext = ftpItmListM;
                 DirFileInfo.Visibility = Visibility.Visible;
+                spFileRename.Visibility = Visibility.Collapsed;
+                spDirectoryRename.Visibility = Visibility.Collapsed;
                 textBlock.Cursor = Cursors.Arrow;
             }
             catch (Exception ex)
             {
+                ErrorHandler.Error(ex);
                 Console.WriteLine(ex);
                 DirFileInfo.Visibility = Visibility.Hidden;
             }
@@ -153,15 +157,16 @@ namespace RetroGameHandler.Views
             var button = (Button)sender;
             var par = button.Parent;
             var ftpItmListM = (FtpListItemModel)button.DataContext;
-            if (MessageBox.Show($"Delete file {ftpItmListM.Name}?", "Delete", MessageBoxButton.YesNo, MessageBoxImage.Question) == MessageBoxResult.No)
+            if (MessageBox.Show($"You are about to delete {(button.Name == "DeleteFile" ? "file" : "directory")} {ftpItmListM.Name} in {ftpItmListM.ParentPath}!\nThis can damage the os or make applications unusable!\nSo if you are not sure what you are doing, please do not proceed!\nDo you like to proceed?", "Delete", MessageBoxButton.YesNo, MessageBoxImage.Warning) == MessageBoxResult.No)
             {
                 return;
             }
 
             var ftpHelper = new FtpHelper();
             var cancellationToken = new CancellationToken();
-            await ftpHelper.DeletedFileAsync(ftpItmListM.FullName, cancellationToken);
-            MessageBox.Show("File deleted");
+            if (button.Name == "DeleteFile") await ftpHelper.DeletedFileAsync(ftpItmListM.FullName, cancellationToken);
+            else await ftpHelper.DeleteDirectoryAsync(ftpItmListM.FullName, cancellationToken);
+            MessageBox.Show($"{(button.Name == "DeleteFile" ? "File" : "Directory")} {ftpItmListM.Name} deleted");
             var holeDamThing = (StartViewModel)FileList.DataContext;
             var path = ftpItmListM.FullName;
             var dt = await holeDamThing.FindItemByPathAsync(ftpItmListM.ParentPath);
@@ -230,5 +235,127 @@ namespace RetroGameHandler.Views
             notPad.LoadTextPath(ftpItmListM.FullName, ftpItmListM.Name);
             PageHandler.SelectedPage<NotePad>();
         }
+
+        private void TextBox_PreviewMouseLeftButtonDown(object sender, MouseButtonEventArgs e)
+        {
+            var textBox = (TextBox)sender;
+            textBox.SelectAll();
+            textBox.Focus();
+        }
+
+        private void btnCopyPath_Click(object sender, RoutedEventArgs e)
+        {
+            var button = (Button)sender;
+            var par = button.Parent;
+            var ftpItmListM = (FtpListItemModel)button.DataContext;
+            Clipboard.SetText(ftpItmListM.FullName);
+        }
+
+        private void RenameFolder_Click(object sender, RoutedEventArgs e)
+        {
+            spDirectoryRename.Visibility = Visibility.Visible;
+        }
+
+        private void DirectoryRenameClose_Click(object sender, RoutedEventArgs e)
+        {
+            spDirectoryRename.Visibility = Visibility.Collapsed;
+        }
+
+        private async void DirectorySaveRename_Click(object sender, RoutedEventArgs e)
+        {
+            var button = (Button)sender;
+            var par = button.Parent;
+            var ftpItmListM = (FtpListItemModel)button.DataContext;
+            var oldDirectory = ftpItmListM.mFullName;
+            var newName = @"/" + ftpItmListM.Name;
+            var newDirectoryName = ftpItmListM.ParentPath + newName;
+            if (MessageBox.Show($"You are about to rename {oldDirectory} to {newDirectoryName}\nThis can damage the os or make applications unusable!\nSo if you are not sure what you are doing, please do not proceed!\nDo you like to proceed?", "Rename Directory?", MessageBoxButton.YesNo, MessageBoxImage.Warning) == MessageBoxResult.No) return;
+            spDirectoryRename.Visibility = Visibility.Collapsed;
+            var ftpHelper = new FtpHelper();
+            var cancellationToken = new CancellationToken();
+            await ftpHelper.RenameDirectory(ftpItmListM.mFullName, newDirectoryName, cancellationToken);
+            MessageBox.Show($"Rename directory ok");
+            var holeDamThing = (StartViewModel)FileList.DataContext;
+            var dt = await holeDamThing.FindItemByPathAsync(ftpItmListM.ParentPath);
+            if (dt != null)
+            {
+                await holeDamThing.GetFtpListItems(dt.Items, dt.FullName, true);
+            }
+        }
+
+        private void FileRenameClose_Click(object sender, RoutedEventArgs e)
+        {
+            spFileRename.Visibility = Visibility.Collapsed;
+        }
+
+        private async void FileSaveRename_Click(object sender, RoutedEventArgs e)
+        {
+            var button = (Button)sender;
+            var par = button.Parent;
+            var ftpItmListM = (FtpListItemModel)button.DataContext;
+            var oldFilePath = ftpItmListM.mFullName;
+            var newName = @"/" + ftpItmListM.Name;
+            var newFileFullName = ftpItmListM.ParentPath + newName;
+            if (MessageBox.Show($"You are about to rename {oldFilePath} to {newFileFullName}\nThis can damage the os or make applications unusable!\nSo if you are not sure what you are doing, please do not proceed!\nDo you like to proceed?", "Rename File?", MessageBoxButton.YesNo, MessageBoxImage.Warning) == MessageBoxResult.No) return;
+            spFileRename.Visibility = Visibility.Collapsed;
+            var ftpHelper = new FtpHelper();
+            var cancellationToken = new CancellationToken();
+            await ftpHelper.RenameFile(ftpItmListM.mFullName, newFileFullName, cancellationToken);
+            MessageBox.Show($"Rename File ok");
+            var holeDamThing = (StartViewModel)FileList.DataContext;
+            var dt = await holeDamThing.FindItemByPathAsync(ftpItmListM.ParentPath);
+            if (dt != null)
+            {
+                await holeDamThing.GetFtpListItems(dt.Items, dt.FullName, true);
+            }
+        }
+
+        private void RenameFile_Click(object sender, RoutedEventArgs e)
+        {
+            spFileRename.Visibility = Visibility.Visible;
+        }
+
+        private async void FileList_SelectedItemChanged_1(object sender, RoutedPropertyChangedEventArgs<object> e)
+        {
+            try
+            {
+                var treeView = (TreeView)sender;
+                var textBlock = treeView.SelectedItem;
+
+                treeView.Cursor = Cursors.Wait;
+                var ftpItmListM = (FtpListItemModel)textBlock;
+                if (ftpItmListM.IsImage)
+                {
+                    ftpItmListM.GetImage();
+                }
+                if (ftpItmListM.IsOpk)
+                {
+                    var info = await ftpItmListM.GetOpkInfo();
+                    this.Dispatcher.Invoke((Action)delegate // <--- HERE
+                    {
+                        opkName.Text = info?.Name ?? "";
+                        opkComment.Text = info?.Comment ?? "";
+                        opkType.Text = info?.Type ?? "";
+                        opkImage.Source = info?.Image;
+                    });
+                }
+                DirFileInfo.DataContext = ftpItmListM;
+                DirFileInfo.Visibility = Visibility.Visible;
+                treeView.Cursor = Cursors.Arrow;
+                spFileRename.Visibility = Visibility.Collapsed;
+                spDirectoryRename.Visibility = Visibility.Collapsed;
+            }
+            catch (Exception ex)
+            {
+                ErrorHandler.Error(ex);
+                Console.WriteLine(ex);
+                DirFileInfo.Visibility = Visibility.Hidden;
+            }
+        }
+
+        //private string ReplaceLastOccurrence(string str, string toReplace, string replacement)
+        //{
+        //    return Regex.Replace(str, $@"^(.*){toReplace}(.*?)$", $"$1{replacement}$2");
+        //}
     }
 }
